src/requirements.txt
black==23.3.0
celery==5.2.7
Django==4.2.1
django-celery-beat==2.5.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
django-rest-swagger==2.2.0
django-filter==23.2
ipython==8.13.2
jedi==0.18.2
marshmallow==3.19.0
numpy==1.24.3
Pillow==9.5.0
pre-commit==3.3.1
# psycopg2==2.9.6
python-json-logger==2.0.7
redis==4.5.5
requests==2.30.0